Revert "Fix cooldown data"
This reverts commit 60abca71bdc91bbb63a6782621dd07b41b4fec1b.

+++ XIVSlothCombo/Combos/PvE/BRD.cs
@@ -654,7 +654,7 @@ namespace XIVSlothCombo.Combos.PvE
 
                         if (LevelChecked(Bloodletter) && ((!openerFinished && IsOnCooldown(RagingStrikes)) || openerFinished))
                         {
-                            uint bloodletterCharges = GetRemainingCharges(Bloodletter);
+                            ushort bloodletterCharges = GetRemainingCharges(Bloodletter);
 
                             if (IsEnabled(CustomComboPreset.BRD_Simple_Pooling) && LevelChecked(WanderersMinuet))
                             {
